add confirmEmail method to user model
